Pull GitHub oAuth out of 'server/server.js'
In order to introduce 100% code coverage to the 'server.js', it is
important to pull out the code that can not be included in the code
coverage.  In this case, it is the GitHub auth code that is to be
pulled out into its own file.
